requirements.txt
Package         Version
--------------- -------
click           8.0.3  
Flask           2.0.2  
itsdangerous    2.0.1  
Jinja2          3.0.3  
MarkupSafe      2.0.1  
pip             20.0.2 
pkg-resources   0.0.0  
psycopg2-binary 2.9.3  
python-dotenv   0.19.2 
setuptools      44.0.0 
Werkzeug        2.0.2  
